fix: move packages to package folder
