+++ cspell.yml
@@ -20,3 +20,6 @@ words:
   - cocogitto
   - juste
   - oknozor
+  - tuto
+  - rspack
+  - vbump
