+++ packages/core/templates.ts
