+++ packages/core/utils.test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from '@std/assert'
-import config from '../readma.config.ts'
+import config from '../../readma.config.ts'
 import { getReadmaConfig } from './utils.ts'
 
 Deno.test('getReadmaConfig', async () => {
