+++ packages/pkg-json/mod.ts
